Reorganize accrual computers into separate files
The various accrual computer implementations were becoming unwieldy
inside one source file.
