+++ src/neuralnet/tally.cpp
@@ -1,5 +1,7 @@
-#include "beacon.h"
 #include "main.h"
+#include "neuralnet/accrual/newbie.h"
+#include "neuralnet/accrual/null.h"
+#include "neuralnet/accrual/research_age.h"
 #include "neuralnet/cpid.h"
 #include "neuralnet/quorum.h"
 #include "neuralnet/researcher.h"
@@ -18,12 +20,6 @@ namespace {
 constexpr size_t TALLY_DAYS = 14;
 
 //!
-//! \brief Used to convert elapsed accrual seconds into days, so we declare it
-//! as a double type.
-//!
-constexpr double ONE_DAY_IN_SECONDS = 86400.0;
-
-//!
 //! \brief Round a value to intervals of 0.025.
 //!
 //! Used to calculate the network magnitude unit.
@@ -33,36 +29,6 @@ double SnapToGrid(double d)
     double dDither = .04;
     double dOut = RoundFromString(RoundToString(d * dDither, 3), 3) / dDither;
     return dOut;
-}
-
-//!
-//! \brief Get the multiplier used to calculate the maximum research reward
-//! amount.
-//!
-//! Survey Results: Start inflation rate: 9%, end=1%, 30 day steps, 9 steps,
-//! mag multiplier start: 2, mag end .3, 9 steps
-//!
-//! \return A value in units of GRC (not COIN) that represents the maximum
-//! research reward expected per block.
-//!
-int64_t GetMaxResearchSubsidy(const int64_t nTime)
-{
-    // Gridcoin Global Daily Maximum Researcher Subsidy Schedule
-    int MaxSubsidy = 500;
-
-    if (nTime > 1447977700)                              MaxSubsidy =  50; // from  11-20-2015 forever
-    else if (nTime >= 1445309276 && nTime <= 1447977700) MaxSubsidy =  75; // between 10-20-2015 and 11-20-2015
-    else if (nTime >= 1438310876 && nTime <= 1445309276) MaxSubsidy = 100; // between 08-01-2015 and 10-20-2015
-    else if (nTime >= 1430352000 && nTime <= 1438310876) MaxSubsidy = 150; // between 05-01-2015 and 07-31-2015
-    else if (nTime >= 1427673600 && nTime <= 1430352000) MaxSubsidy = 200; // between 03-30-2015 and 04-30-2015
-    else if (nTime >= 1425254400 && nTime <= 1427673600) MaxSubsidy = 250; // between 02-28-2015 and 03-30-2015
-    else if (nTime >= 1422576000 && nTime <= 1425254400) MaxSubsidy = 300; // between 01-30-2015 and 02-28-2015
-    else if (nTime >= 1419897600 && nTime <= 1422576000) MaxSubsidy = 400; // between 12-30-2014 and 01-30-2015
-    else if (nTime >= 1417305600 && nTime <= 1419897600) MaxSubsidy = 400; // between 11-30-2014 and 12-30-2014
-    else if (nTime >= 1410393600 && nTime <= 1417305600) MaxSubsidy = 500; // between inception  and 11-30-2014
-
-    // The .5 allows for fractional amounts after the 4th decimal place (used to store the POR indicator)
-    return MaxSubsidy+.5;
 }
 
 //!
@@ -298,355 +264,10 @@ private:
     SuperblockPtr m_current_superblock = std::make_shared<Superblock>();
 }; // ResearcherTally
 
-//!
-//! \brief An implementation of IAccrualComputer that always returns zeros.
-//!
-//! Useful for situations where we need to satisfy the interface but cannot
-//! possibly calculate an accrual.
-//!
-class NullAccrualComputer : public IAccrualComputer
-{
-    // See IAccrualComputer for inherited API documentation.
-
-public:
-    double MagnitudeUnit() const override
-    {
-        return 0.0;
-    }
-
-    int64_t AccrualAge(const ResearchAccount& account) const override
-    {
-        return 0;
-    }
-
-    double AccrualDays(const ResearchAccount& account) const override
-    {
-        return 0.0;
-    }
-
-    int64_t AccrualBlockSpan(const ResearchAccount& account) const override
-    {
-        return 0;
-    }
-
-    double AverageMagnitude(const ResearchAccount& account) const override
-    {
-        return 0.0;
-    }
-
-    int64_t PaymentPerDay(const ResearchAccount& account) const override
-    {
-        return 0;
-    }
-
-    int64_t PaymentPerDayLimit(const ResearchAccount& account) const override
-    {
-        return 0;
-    }
-
-    bool ExceededRecentPayments(const ResearchAccount& account) const override
-    {
-        return false;
-    }
-
-    int64_t ExpectedDaily(const ResearchAccount& account) const override
-    {
-        return 0;
-    }
-
-    int64_t RawAccrual(const ResearchAccount& account) const override
-    {
-        return 0;
-    }
-
-    int64_t Accrual(const ResearchAccount& account) const override
-    {
-        return 0;
-    }
-}; // NullAccrualComputer
-
-//!
-//! \brief An accrual calculator for a CPID that never earned a research reward
-//! before.
-//!
-class NewbieAccrualComputer : public NullAccrualComputer
-{
-    // See IAccrualComputer for inherited API documentation.
-
-public:
-    //!
-    //! \brief Initialze an accrual calculator for a CPID that never earned
-    //! a research reward before.
-    //!
-    //! \param cpid           CPID to calculate research accrual for.
-    //! \param payment_time   Time of payment to calculate rewards at.
-    //! \param magnitude_unit Current network magnitude unit to factor in.
-    //!
-    NewbieAccrualComputer(
-        const Cpid cpid,
-        const int64_t payment_time,
-        const double magnitude_unit)
-        : m_cpid(cpid)
-        , m_payment_time(payment_time)
-        , m_magnitude_unit(magnitude_unit)
-    {
-    }
-
-    double MagnitudeUnit() const override
-    {
-        return m_magnitude_unit;
-    }
-
-    int64_t AccrualAge(const ResearchAccount& account) const override
-    {
-        const int64_t beacon_time = BeaconTimeStamp(m_cpid.ToString());
-
-        if (beacon_time == 0) {
-            return 0;
-        }
-
-        return m_payment_time - beacon_time;
-    }
-
-    double AccrualDays(const ResearchAccount& account) const override
-    {
-        return AccrualAge(account) / ONE_DAY_IN_SECONDS;
-    }
-
-    double AverageMagnitude(const ResearchAccount& account) const override
-    {
-        return account.m_magnitude;
-    }
-
-    int64_t PaymentPerDayLimit(const ResearchAccount& account) const override
-    {
-        return 500 * COIN;
-    }
-
-    bool ExceededRecentPayments(const ResearchAccount& account) const override
-    {
-        return RawAccrual(account) > PaymentPerDayLimit(account) * COIN;
-    }
-
-    int64_t ExpectedDaily(const ResearchAccount& account) const override
-    {
-        return account.m_magnitude * m_magnitude_unit * COIN;
-    }
-
-    int64_t RawAccrual(const ResearchAccount& account) const override
-    {
-        return AccrualDays(account) * ExpectedDaily(account);
-    }
-
-    int64_t Accrual(const ResearchAccount& account) const override
-    {
-        constexpr int64_t six_months = ONE_DAY_IN_SECONDS * 30 * 6;
-
-        if (AccrualAge(account) >= six_months) {
-            if (account.m_magnitude <= 0) {
-                return 0;
-            }
-
-            if (fDebug) {
-                LogPrintf(
-                    "Accrual: %s Invalid Beacon, Using 0.01 age bootstrap",
-                    m_cpid.ToString());
-            }
-
-            return ((double)account.m_magnitude / 100) * COIN + (1 * COIN);
-        }
-
-        const int64_t accrual = RawAccrual(account);
-
-        if (accrual > 500 * COIN) {
-            if (fDebug) {
-                LogPrintf(
-                    "Accrual: %s Newbie special stake capped to 500 GRC.",
-                    m_cpid.ToString());
-            }
-
-            return 500 * COIN;
-        }
-
-        return accrual + (1 * COIN);
-    }
-
-private:
-    const Cpid m_cpid;             //!< CPID to calculate research accrual for.
-    const int64_t m_payment_time;  //!< Time of payment to calculate rewards at.
-    const double m_magnitude_unit; //!< Network magnitude unit to factor in.
-}; // NewbieAccrualComputer
-
 ResearcherTally g_researcher_tally; //!< Tracks lifetime research rewards.
 NetworkTally g_network_tally;       //!< Tracks two-week network averages.
 
 } // Anonymous namespace
-
-// -----------------------------------------------------------------------------
-// Class: ResearchAgeComputer
-// -----------------------------------------------------------------------------
-
-ResearchAgeComputer::ResearchAgeComputer(
-    const Cpid cpid,
-    const int64_t payment_time,
-    const double magnitude_unit,
-    const uint32_t last_height)
-    : m_cpid(cpid)
-    , m_payment_time(payment_time)
-    , m_magnitude_unit(magnitude_unit)
-    , m_last_height(last_height)
-{
-}
-
-int64_t ResearchAgeComputer::MaxReward(const int64_t payment_time)
-{
-    return GetMaxResearchSubsidy(payment_time) * 255 * COIN;
-}
-
-double ResearchAgeComputer::MagnitudeUnit() const
-{
-    return m_magnitude_unit;
-}
-
-int64_t ResearchAgeComputer::AccrualAge(const ResearchAccount& account) const
-{
-    if (const BlockPtrOption pindex_option = account.LastRewardBlock()) {
-        const CBlockIndex* const pindex = *pindex_option;
-
-        if (m_payment_time > pindex->nTime) {
-            return m_payment_time - pindex->nTime;
-        }
-    }
-
-    return 0;
-}
-
-int64_t ResearchAgeComputer::AccrualBlockSpan(const ResearchAccount& account) const
-{
-    return m_last_height - account.LastRewardHeight();
-}
-
-double ResearchAgeComputer::AccrualDays(const ResearchAccount& account) const
-{
-    return AccrualAge(account) / ONE_DAY_IN_SECONDS;
-}
-
-double ResearchAgeComputer::AverageMagnitude(const ResearchAccount& account) const
-{
-    const uint16_t last_magnitude = account.LastRewardMagnitude();
-
-    if (m_last_height - account.LastRewardHeight() <= BLOCKS_PER_DAY * 20) {
-        return (double)(last_magnitude + account.m_magnitude) / 2;
-    }
-
-    // ResearchAge: If the accrual age is more than 20 days, add in the midpoint
-    // lifetime average magnitude to ensure the overall avg magnitude accurate:
-    //
-    // TODO: use superblock windows to calculate more precise average
-    //
-    const double total_mag = last_magnitude
-        + account.AverageLifetimeMagnitude()
-        + account.m_magnitude;
-
-    return total_mag / 3;
-}
-
-int64_t ResearchAgeComputer::PaymentPerDay(const ResearchAccount& account) const
-{
-    if (account.IsNew()) {
-        return 0;
-    }
-
-    const int64_t elapsed = m_payment_time - account.FirstRewardTime();
-    const double lifetime_days = elapsed / ONE_DAY_IN_SECONDS;
-
-    // The extra 0.01 days was used in old code as a lazy way to avoid division
-    // by zero errors. The payment per day limit exception in historical blocks
-    // depends on this extra value, so we keep it. This especially manifests on
-    // testnet where CPIDs stake frequently enough to activate the rule:
-    //
-    return account.m_total_research_subsidy / (lifetime_days + 0.01);
-}
-
-int64_t ResearchAgeComputer::PaymentPerDayLimit(const ResearchAccount& account) const
-{
-    return account.AverageLifetimeMagnitude() * m_magnitude_unit * COIN * 5;
-}
-
-bool ResearchAgeComputer::ExceededRecentPayments(const ResearchAccount& account) const
-{
-    return PaymentPerDay(account) > PaymentPerDayLimit(account);
-}
-
-int64_t ResearchAgeComputer::ExpectedDaily(const ResearchAccount& account) const
-{
-    return account.m_magnitude * m_magnitude_unit * COIN;
-}
-
-int64_t ResearchAgeComputer::RawAccrual(const ResearchAccount& account) const
-{
-    double current_avg_magnitude = AverageMagnitude(account);
-
-    // Legacy sanity check for unlikely magnitude values:
-    if (current_avg_magnitude > 20000) {
-        current_avg_magnitude = 20000;
-    }
-
-    return AccrualDays(account) * current_avg_magnitude * m_magnitude_unit * COIN;
-}
-
-int64_t ResearchAgeComputer::Accrual(const ResearchAccount& account) const
-{
-    // Note that if the RA block span < 10, we want to return 0 for the accrual
-    // amount so the CPID can still receive an accurate accrual in the future:
-    //
-    if (AccrualBlockSpan(account) < 10) {
-        if (fDebug) {
-            LogPrintf(
-                "Accrual: %s Block Span less than 10 (%d) -> Accrual 0 (would be %s)",
-                m_cpid.ToString(),
-                AccrualBlockSpan(account),
-                FormatMoney(RawAccrual(account)));
-        }
-
-        return 0;
-    }
-
-    // Since this condition can occur when a user ramps up their computing
-    // power, let's return 0 to avoid shortchanging a researcher. Instead,
-    // owed GRC continues to accrue and will be paid later after payments-
-    // per-day falls below 5:
-    //
-    if (ExceededRecentPayments(account)) {
-        if (fDebug) {
-            LogPrintf("Accrual: %s RA-PPD, "
-                "PPD=%s, "
-                "unit=%f, "
-                "RAAvgMag=%f, "
-                "RASubsidy=%s, "
-                "RALowLockTime=%" PRIu64 " "
-                "-> Accrual 0 (would be %s)",
-                m_cpid.ToString(),
-                FormatMoney(PaymentPerDay(account)),
-                m_magnitude_unit,
-                account.AverageLifetimeMagnitude(),
-                FormatMoney(account.m_total_research_subsidy),
-                account.FirstRewardTime(),
-                FormatMoney(RawAccrual(account)));
-        }
-
-        return 0;
-    }
-
-    const int64_t accrual = RawAccrual(account);
-    const int64_t max_reward = MaxReward(m_payment_time);
-
-    if (accrual > max_reward) {
-        return max_reward;
-    }
-
-    return accrual;
-}
 
 // -----------------------------------------------------------------------------
 // Class: Tally
