+++ src/neuralnet/tally.h
@@ -1,217 +1,13 @@
 #pragma once
 
 #include "neuralnet/account.h"
+#include "neuralnet/accrual/computer.h"
 
 class CBlockIndex;
 
 namespace NN {
 
 class Cpid;
-
-//!
-//! \brief A calculator that computes the accrued research rewards for a
-//! research account.
-//!
-class IAccrualComputer
-{
-public:
-    //!
-    //! \brief Destructor.
-    //!
-    virtual ~IAccrualComputer() { }
-
-    //!
-    //! \brief Get the magnitude unit factored into the reward calculation.
-    //!
-    virtual double MagnitudeUnit() const = 0;
-
-    //!
-    //! \brief Get the time elapsed since the account's last research reward.
-    //!
-    //! \return Elapsed time in seconds.
-    //!
-    virtual int64_t AccrualAge(const ResearchAccount& account) const = 0;
-
-    //!
-    //! \brief Get the number of days since the account's last research reward.
-    //!
-    virtual double AccrualDays(const ResearchAccount& account) const = 0;
-
-    //!
-    //! \brief Get the number of blocks since the account's last research reward.
-    //!
-    virtual int64_t AccrualBlockSpan(const ResearchAccount& account) const = 0;
-
-    //!
-    //! \brief Get the average magnitude of the account over the accrual time
-    //! span.
-    //!
-    virtual double AverageMagnitude(const ResearchAccount& account) const = 0;
-
-    //!
-    //! \brief Get the average daily research payment over the lifetime of the
-    //! account.
-    //!
-    //! \return Average research payment in units of 1/100000000 GRC.
-    //!
-    virtual int64_t PaymentPerDay(const ResearchAccount& account) const = 0;
-
-    //!
-    //! \brief Get the average daily research payment limit of the account.
-    //!
-    //! \return Payment per day limit in units of 1/100000000 GRC.
-    //!
-    virtual int64_t PaymentPerDayLimit(const ResearchAccount& account) const = 0;
-
-    //!
-    //! \brief Determine whether the account exceeded the daily payment limit.
-    //!
-    //! \return \c true if the average daily research payment amount exceeds
-    //! the calculated daily payment limit of the account.
-    //!
-    virtual bool ExceededRecentPayments(const ResearchAccount& account) const = 0;
-
-    //!
-    //! \brief Get the expected daily research payment for the account based on
-    //! the current network payment conditions.
-    //!
-    //! \return Expected daily payment in units of 1/100000000 GRC.
-    //!
-    virtual int64_t ExpectedDaily(const ResearchAccount& account) const = 0;
-
-    //!
-    //! \brief Get the pending research reward for the account without applying
-    //! any limit rules.
-    //!
-    //! \return Pending payment in units of 1/100000000 GRC.
-    //!
-    virtual int64_t RawAccrual(const ResearchAccount& account) const = 0;
-
-    //!
-    //! \brief Get the pending research reward for the account as expected by
-    //! the network.
-    //!
-    //! \return Pending payment in units of 1/100000000 GRC.
-    //!
-    virtual int64_t Accrual(const ResearchAccount& account) const = 0;
-};
-
-//!
-//! \brief Wraps an object that implements the IAccrualComputer interface.
-//!
-typedef std::unique_ptr<IAccrualComputer> AccrualComputer;
-
-//!
-//! \brief A calculator that computes the accrued research rewards for a
-//! research account using research age rules.
-//!
-class ResearchAgeComputer : public IAccrualComputer
-{
-public:
-    //!
-    //! \brief Initialze a research age accrual calculator.
-    //!
-    //! \param cpid           CPID to calculate research accrual for.
-    //! \param payment_time   Time of payment to calculate rewards at.
-    //! \param magnitude_unit Current network magnitude unit to factor in.
-    //! \param last_height    Height of the block for the reward.
-    //!
-    ResearchAgeComputer(
-        const Cpid cpid,
-        const int64_t payment_time,
-        const double magnitude_unit,
-        const uint32_t last_height);
-
-    //!
-    //! \brief Get the maximum research reward payable in one block.
-    //!
-    //! \param payment_time Time of payment to calculate rewards for.
-    //!
-    //! \return Max reward allowed in units of 1/100000000 GRC.
-    //!
-    static int64_t MaxReward(const int64_t payment_time);
-
-    //!
-    //! \brief Get the magnitude unit factored into the reward calculation.
-    //!
-    double MagnitudeUnit() const override;
-
-    //!
-    //! \brief Get the time elapsed since the account's last research reward.
-    //!
-    //! \return Elapsed time in seconds.
-    //!
-    int64_t AccrualAge(const ResearchAccount& account) const override;
-
-    //!
-    //! \brief Get the number of days since the account's last research reward.
-    //!
-    double AccrualDays(const ResearchAccount& account) const override;
-
-    //!
-    //! \brief Get the number of blocks since the account's last research reward.
-    //!
-    int64_t AccrualBlockSpan(const ResearchAccount& account) const override;
-
-    //!
-    //! \brief Get the average magnitude of the account over the accrual time
-    //! span.
-    //!
-    double AverageMagnitude(const ResearchAccount& account) const override;
-
-    //!
-    //! \brief Get the average daily research payment over the lifetime of the
-    //! account.
-    //!
-    //! \return Average research payment in units of 1/100000000 GRC.
-    //!
-    int64_t PaymentPerDay(const ResearchAccount& account) const override;
-
-    //!
-    //! \brief Get the average daily research payment limit of the account.
-    //!
-    //! \return Payment per day limit in units of 1/100000000 GRC.
-    //!
-    int64_t PaymentPerDayLimit(const ResearchAccount& account) const override;
-
-    //!
-    //! \brief Determine whether the account exceeded the daily payment limit.
-    //!
-    //! \return \c true if the average daily research payment amount exceeds
-    //! the calculated daily payment limit of the account.
-    //!
-    bool ExceededRecentPayments(const ResearchAccount& account) const override;
-
-    //!
-    //! \brief Get the expected daily research payment for the account based on
-    //! the current network payment conditions.
-    //!
-    //! \return Expected daily payment in units of 1/100000000 GRC.
-    //!
-    int64_t ExpectedDaily(const ResearchAccount& account) const override;
-
-    //!
-    //! \brief Get the pending research reward for the account without applying
-    //! any limit rules.
-    //!
-    //! \return Pending payment in units of 1/100000000 GRC.
-    //!
-    int64_t RawAccrual(const ResearchAccount& account) const override;
-
-    //!
-    //! \brief Get the pending research reward for the account as expected by
-    //! the network.
-    //!
-    //! \return Pending payment in units of 1/100000000 GRC.
-    //!
-    int64_t Accrual(const ResearchAccount& account) const override;
-
-private:
-    const Cpid m_cpid;             //!< CPID to calculate research accrual for.
-    const int64_t m_payment_time;  //!< Time of payment to calculate rewards at.
-    const double m_magnitude_unit; //!< Network magnitude unit to factor in.
-    const uint32_t m_last_height;  //!< Height of the block for the reward.
-};
 
 //!
 //! \brief The core Gridcoin tally system that processes magnitudes and reward
